+++ common/src/main/java/com/github/yajatkaul/mega_showdown/api/codec/SoloFusion.java
@@ -105,7 +105,7 @@ public record SoloFusion(
                 stack.set(DataComponents.CUSTOM_NAME, Component.translatable("item.mega_showdown." + namespace + ".inactive"));
             } else if (pokemonStored != null && isMain) {
                 if (aspect_conditions.validate_apply(pokemon)) {
-                    Effect.getEffect(effect.get()).revertEffects(pokemon, aspect_conditions.aspectApply().aspects(), aspect_conditions.aspectApply().pokemonProperties(), null);
+                    Effect.getEffect(effect.get()).applyEffects(pokemon, aspect_conditions.aspectApply().aspects(), aspect_conditions.aspectApply().pokemonProperties(), null);
                 } else {
                     return InteractionResultHolder.pass(stack);
                 }
@@ -118,13 +118,14 @@ public record SoloFusion(
                 stack.remove(MegaShowdownDataComponents.POKEMON_STORAGE.get());
                 stack.set(DataComponents.CUSTOM_NAME, Component.translatable("item.mega_showdown." + namespace + ".inactive"));
             } else if (pokemonStored == null && isFusion) {
+                pokemon.getPersistentData().putUUID("lastOwner", player.getUUID());
                 stack.set(MegaShowdownDataComponents.POKEMON_STORAGE.get(), pokemonStorge.save(registryAccess, pokemon));
                 stack.set(DataComponents.CUSTOM_NAME, Component.translatable("item.mega_showdown." + namespace + ".charged"));
 
                 playerPartyStore.remove(pokemon);
             }
         } else if (pokemonStored != null) {
-            if (!pokemonStored.getTradeable() && pokemonStored.getOwnerPlayer() != player) {
+            if (!pokemonStored.getTradeable() && !pokemonStored.getPersistentData().getUUID("lastOwner").equals(player.getUUID())) {
                 player.displayClientMessage(Component.translatable("message.mega_showdown.untradable")
                         .withStyle(ChatFormatting.RED), true);
                 return InteractionResultHolder.pass(stack);