+++ common/src/main/java/com/github/yajatkaul/mega_showdown/item/custom/fusion/SoloFusion.java
@@ -108,7 +108,7 @@ public class SoloFusion extends ToolTipItem {
                 stack.remove(MegaShowdownDataComponents.POKEMON_STORAGE.get());
                 stack.set(DataComponents.CUSTOM_NAME, Component.translatable("item.mega_showdown." + namespace + ".inactive"));
             } else if (pokemonStored != null && isMain) {
-                Effect.getEffect(effectId).revertEffects(pokemon, applyAspect, Optional.empty(), null);
+                Effect.getEffect(effectId).applyEffects(pokemon, applyAspect, Optional.empty(), null);
                 pokemon.setTradeable(false);
 
                 CompoundTag otherPokemonNbt = pokemonStored.saveToNBT(level.registryAccess(), new CompoundTag());
@@ -117,13 +117,14 @@ public class SoloFusion extends ToolTipItem {
                 stack.remove(MegaShowdownDataComponents.POKEMON_STORAGE.get());
                 stack.set(DataComponents.CUSTOM_NAME, Component.translatable("item.mega_showdown." + namespace + ".inactive"));
             } else if (pokemonStored == null && isFusion) {
+                pokemon.getPersistentData().putUUID("lastOwner", player.getUUID());
                 stack.set(MegaShowdownDataComponents.POKEMON_STORAGE.get(), pokemonStorge.save(registryAccess, pokemon));
                 stack.set(DataComponents.CUSTOM_NAME, Component.translatable("item.mega_showdown." + namespace + ".charged"));
 
                 playerPartyStore.remove(pokemon);
             }
         } else if (pokemonStored != null) {
-            if (!pokemonStored.getTradeable() && pokemonStored.getOwnerPlayer() != player) {
+            if (!pokemonStored.getTradeable() && !pokemonStored.getPersistentData().getUUID("lastOwner").equals(player.getUUID())) {
                 player.displayClientMessage(Component.translatable("message.mega_showdown.untradable")
                         .withStyle(ChatFormatting.RED), true);
                 return InteractionResultHolder.pass(stack);