+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # 1.0+1.8.0-release
-- Added ZA Mega's
+- Added some ZA Mega's with known abilities
 - Fixed some species additions
 - 1.8 Compatibility
 - Fixed untradable necrozmas being traded using Lunarizers
+- Fixed solo fusion items apply effect not working